requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
httpx==0.27.0
python-dotenv==1.0.1
deta==1.1.0
